+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- add `Array.deduplicate` and `Array.deduplicateByKey` ([#2](https://github.com/seaofvoices/luau-disk/pull/2))
 - add `Array.findMap` and `Array.popFirst` ([#1](https://github.com/seaofvoices/luau-disk/pull/1))
 
 ## 0.1.0